got ManageStudents screen to render with graphql, but mutations and local state still need to be migrated
