+++ apps/src/templates/manageStudents/ManageStudents.jsx
@@ -1,52 +1,79 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {connect} from 'react-redux';
+import {OAuthSectionTypes} from '@cdo/apps/lib/ui/accounts/constants';
 import ManageStudentsTable from './ManageStudentsTable';
 import SyncOmniAuthSectionControl from '@cdo/apps/lib/ui/SyncOmniAuthSectionControl';
-import {loadSectionStudentData} from '@cdo/apps/templates/manageStudents/manageStudentsRedux';
 import Spinner from '@cdo/apps/code-studio/pd/components/spinner';
 
-class ManageStudents extends React.Component {
-  static propTypes = {
-    studioUrlPrefix: PropTypes.string,
+import {gql, useQuery} from '@apollo/client';
+import {
+  blankStudentTransfer,
+  blankStudentTransferStatus
+} from './manageStudentsTypes';
 
-    // Provided by redux
-    sectionId: PropTypes.number,
-    isLoadingStudents: PropTypes.bool.isRequired,
-    loadSectionStudentData: PropTypes.func.isRequired
-  };
-
-  componentDidMount() {
-    this.props.loadSectionStudentData(this.props.sectionId);
-  }
-
-  render() {
-    const {sectionId, studioUrlPrefix, isLoadingStudents} = this.props;
-
-    return (
-      <div>
-        {isLoadingStudents && <Spinner />}
-        {!isLoadingStudents && (
-          <div>
-            <SyncOmniAuthSectionControl sectionId={sectionId} />
-            <ManageStudentsTable studioUrlPrefix={studioUrlPrefix} />
-          </div>
-        )}
-      </div>
-    );
-  }
-}
-
-export const UnconnectedManageStudents = ManageStudents;
-
-export default connect(
-  state => ({
-    sectionId: state.sectionData.section.id,
-    isLoadingStudents: state.manageStudents.isLoadingStudents
-  }),
-  dispatch => ({
-    loadSectionStudentData(sectionId) {
-      dispatch(loadSectionStudentData(sectionId));
+export const GET_SECTION = gql`
+  query GetSection($sectionId: ID!) {
+    section(id: $sectionId) {
+      id
+      code
+      name
+      loginType
+      students {
+        id
+        name
+        username
+        email
+        age
+        gender
+        secretWords
+        secretPicturePath
+        hasEverSignedIn
+        userType
+      }
     }
-  })
-)(ManageStudents);
+  }
+`;
+
+const ManageStudents = ({sectionId, studioUrlPrefix}) => {
+  const {data, loading, error} = useQuery(GET_SECTION, {
+    variables: {sectionId: sectionId}
+  });
+  if (loading) {
+    return <Spinner />;
+  }
+  if (error) {
+    return <p>ERROR: {error.message}</p>;
+  }
+  if (!data?.section) {
+    return <p>Not found</p>;
+  }
+  const section = data.section;
+
+  return (
+    <React.Fragment>
+      <SyncOmniAuthSectionControl
+        sectionId={sectionId}
+        sectionCode={section.code}
+        sectionName={section.name}
+        sectionProvider={OAuthSectionTypes[section.loginType]}
+      />
+      <ManageStudentsTable
+        studioUrlPrefix={studioUrlPrefix}
+        sectionCode={section.code}
+        sectionName={section.name}
+        studentData={section.students}
+        loginType={section.loginType}
+        editingData={{}}
+        addStatus={{}}
+        transferData={{...blankStudentTransfer}}
+        transferStatus={{...blankStudentTransferStatus}}
+      />
+    </React.Fragment>
+  );
+};
+ManageStudents.propTypes = {
+  sectionId: PropTypes.number.isRequired,
+  studioUrlPrefix: PropTypes.string.isRequired
+};
+
+export default ManageStudents;
