+++ apps/src/templates/manageStudents/PrintLoginCards.jsx
@@ -3,7 +3,7 @@ import React, {Component} from 'react';
 import Button from '@cdo/apps/templates/Button';
 import i18n from '@cdo/locale';
 import firehoseClient from '@cdo/apps/lib/util/firehose';
-import {PrintLoginCardsButtonMetricsCategory} from '@cdo/apps/templates/manageStudents/manageStudentsRedux';
+import {PrintLoginCardsButtonMetricsCategory} from '@cdo/apps/templates/manageStudents/manageStudentsTypes';
 
 export default class PrintLoginCards extends Component {
   static propTypes = {
